· Changed Stream::read_byte return type to std::uint32_t · Added new BinaryReader::read specialization for int8_t

+++ SceneR.Framework/Headers/System/IO/FileStream.hpp
@@ -65,7 +65,7 @@ namespace System
             /**
              * Reads a byte from the stream or returns -1 if at the end of the stream.
              */
-            std::uint8_t read_byte() override;
+            std::uint32_t read_byte() override;
 
             /**
              * Reads a sequence of bytes from the current stream.
